+++ template/src/screens/resetPassword/ForgotPasswordScreen.tsx
@@ -1,78 +1,58 @@
 import {useNavigation} from '@react-navigation/native';
 import type {NativeStackNavigationProp} from '@react-navigation/native-stack';
-import {toLower} from 'lodash';
 import React, {useRef, useState} from 'react';
 import {
-  findNodeHandle,
   NativeSyntheticEvent,
   StyleSheet,
   Text,
   TextInputFocusEventData,
+  findNodeHandle,
 } from 'react-native';
 import {KeyboardAwareScrollView} from 'react-native-keyboard-aware-scroll-view';
 import {ButtonType} from '../../../types';
 import {PrimaryButton} from '../../common/components/PrimaryButton';
 import {PrimaryTextInput} from '../../common/components/PrimaryTextInput';
 import {localization} from '../../common/localization/localization';
-import {emptyValidation} from '../../common/validations/commonValidations';
 import {useInputError} from '../../common/validations/hooks/useInputError';
 import {emailValidations} from '../../common/validations/profileValidations';
 import {useAppDispatch} from '../../core/store/reduxHelpers';
-import {userLogin} from '../../core/store/user/userActions';
+import {resetPassword} from '../../core/store/user/userActions';
 import {Colors} from '../../core/theme/colors';
 import {CommonSizes} from '../../core/theme/commonSizes';
 import {CommonStyles} from '../../core/theme/commonStyles';
 import type {RootStackParamList} from '../../navigation/types';
 
-export function Login(): JSX.Element {
+export default function ForgotPasswordScreen(): JSX.Element {
   const dispatch = useAppDispatch();
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [loading, setLoading] = useState(false);
   const navigation =
     useNavigation<NativeStackNavigationProp<RootStackParamList>>();
+  const [email, setEmail] = useState('');
+  const [loading, setLoading] = useState(false);
+
+  const scroll = useRef<KeyboardAwareScrollView>(null);
 
   const {error: emailError, recheckValue: recheckEmail} = useInputError(
     email,
     emailValidations,
   );
-  const {error: passwordError, recheckValue: recheckPassword} = useInputError(
-    password,
-    emptyValidation,
-  );
 
-  async function loginUser() {
-    const emailValid = recheckEmail() === null;
-    const passwordValid = recheckPassword() === null;
+  function scrollToInput(reactNode: any) {
+    scroll.current?.scrollToFocusedInput(reactNode);
+  }
 
-    if (!emailValid || !passwordValid) {
+  async function handleResetPassword() {
+    const isEmailValid = recheckEmail() === null;
+    if (!isEmailValid) {
       return;
     }
 
     setLoading(true);
-    await dispatch(
-      userLogin({
-        email: toLower(email),
-        password,
-      }),
-    );
+    await dispatch(resetPassword({email: email.toLowerCase()}));
     setLoading(false);
   }
-  const scroll = useRef<KeyboardAwareScrollView>(null);
-  function scrollToInput(reactNode: any) {
-    // Add a 'scroll' ref to your ScrollView
 
-    // setTimeout(() => {
-    scroll.current?.scrollToFocusedInput(reactNode);
-    // }, 500);
-  }
-
-  const goToRegistration = () => {
-    navigation.navigate('Registration');
-  };
-
-  const goToForgotPassword = () => {
-    navigation.navigate('ForgotPassword');
+  const goToLogin = () => {
+    navigation.navigate('Login');
   };
 
   return (
@@ -81,11 +61,17 @@ export function Login(): JSX.Element {
       resetScrollToCoords={{x: 0, y: 0}}
       scrollEnabled={true}
       enableOnAndroid={true}
-      testID={'MainPageID'}
       contentContainerStyle={styles.contentContainer}
       contentInsetAdjustmentBehavior={'automatic'}
       style={styles.container}>
-      <Text style={CommonStyles.h1_semiBold}>{localization.login.Login}</Text>
+      <Text style={CommonStyles.h1_semiBold}>
+        {localization.login.forgotPassword.title}
+      </Text>
+
+      <Text style={styles.description}>
+        {localization.login.forgotPassword.description}
+      </Text>
+
       <PrimaryTextInput
         onFocus={(event: NativeSyntheticEvent<TextInputFocusEventData>) => {
           scrollToInput(findNodeHandle(event.target));
@@ -98,33 +84,17 @@ export function Login(): JSX.Element {
         placeholder={localization.login.EnterEmail}
         error={emailError}
       />
-      <PrimaryTextInput
-        onFocus={(event: NativeSyntheticEvent<TextInputFocusEventData>) => {
-          scrollToInput(findNodeHandle(event.target));
-        }}
-        clearButtonMode="never"
-        value={password}
-        secureTextEntry
-        onChangeText={setPassword}
-        containerStyle={CommonStyles.textInputContainer}
-        label={localization.login.Password}
-        placeholder={localization.login.EnterPassword}
-        error={passwordError}
-      />
-      <PrimaryButton
-        onPress={goToForgotPassword}
-        label={localization.login.forgetPassword}
-        type={ButtonType.borderless}
-      />
+
       <PrimaryButton
         isLoading={loading}
-        onPress={loginUser}
-        label={localization.login.continue}
+        onPress={handleResetPassword}
+        label={localization.login.forgotPassword.resetPassword}
         type={ButtonType.solid}
       />
+
       <PrimaryButton
-        onPress={goToRegistration}
-        label={localization.login.notMember}
+        onPress={goToLogin}
+        label={localization.login.forgotPassword.backToLogin}
         type={ButtonType.borderless}
       />
     </KeyboardAwareScrollView>
@@ -132,10 +102,6 @@ export function Login(): JSX.Element {
 }
 
 const styles = StyleSheet.create({
-  imageContainer: {
-    ...CommonStyles.flex1,
-    justifyContent: 'flex-end',
-  },
   container: {
     flexGrow: 1,
     backgroundColor: Colors.white,
@@ -149,14 +115,12 @@ const styles = StyleSheet.create({
     paddingVertical: 26,
     gap: 16,
   },
-  forgotPassword: {
+  description: {
     ...CommonStyles.normalText,
-    color: Colors.primary100,
-    textAlign: 'right',
-    marginTop: 8,
-    marginBottom: 24,
+    textAlign: 'center',
+    marginBottom: 8,
   },
-  registerLink: {
+  loginLink: {
     ...CommonStyles.normalText,
     color: Colors.primary100,
     textAlign: 'center',